Class InformationCriteria

Class of object returned by the QuadrupenFit$criteria() method or the criteria() function. Owns print() and plot() methods.

Active bindings

data

a data frame containing the values of various information criteria (AIC, BIC, lmBIC, eBIC, GCV) along the value of lambda1.

lambda

vector of \(\lambda_1\) (\(\ell_1\) or \(\ell_\infty\) penalty levels) for which each cross-validation has been performed.

names

a vector of characters storing the names of the precomputed criteria

Methods

Public methods

InformationCriteria$new()

InformationCriteria$show()

InformationCriteria$print()

InformationCriteria$plot()

InformationCriteria$clone()

InformationCriteria$new()

Constructor for a InformationCriteria object Should be called internally by an object QuadrupenFit$criteria()

Usage

InformationCriteria$new(value)

Arguments

value

data frame storing output of QuadrupenFit$criteria()

InformationCriteria$show()

User friendly print method

Usage

InformationCriteria$show()

InformationCriteria$print()

User friendly print method

Usage

InformationCriteria$print()

InformationCriteria$plot()

Plot the the desired criteria

Usage

InformationCriteria$plot(
  criteria = self$names,
  log_scale = TRUE,
  xvar = c("lambda", "fraction", "df"),
  title = "Information Criteria"
)

Arguments

criteria

a vector of character with the criteria to plot. The default plot all the criteria available (stored in the field names)

log_scale

logical; indicates if a log-scale should be used when xvar="lambda". Default is TRUE.

xvar

variable to plot on the X-axis: either "df" (the estimated degrees of freedom), "lambda" (\(\lambda_1\) penalty level) or "fraction" (\(\ell_1\)-norm of the coefficients). Default is set to "lambda".

title

graph title

Returns

a ggplot2 object

InformationCriteria$clone()

The objects of this class are cloneable with this method.

Usage

InformationCriteria$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.