Class "RidgeRegressionFit"

Class "RidgeRegressionFit"

Class "RidgeRegressionFit"

Details

Class of object returned by the fitting function ridge(). Inherits fields and methods of QuadrupenFit.

See also

Super class

quadrupen::QuadrupenFit -> RidgeRegressionFit

Active bindings

penalty

character describing the regularizer/penalty

Methods

Public methods

RidgeRegressionFit$new()

RidgeRegressionFit$clone()

Inherited methods

quadrupen::QuadrupenFit$criteria()

quadrupen::QuadrupenFit$cross_validate()

quadrupen::QuadrupenFit$fit()

quadrupen::QuadrupenFit$get_model()

quadrupen::QuadrupenFit$plot()

quadrupen::QuadrupenFit$plot_path()

quadrupen::QuadrupenFit$predict()

quadrupen::QuadrupenFit$print()

quadrupen::QuadrupenFit$show()

quadrupen::QuadrupenFit$stability()

Method new()

Initialize a RidgeRegressionFit model

Usage

RidgeRegressionFit$new(data, intercept, regParam)

Arguments

data

a DataModel object

intercept

a logical; should an intercept be included in the mode?

regParam

a list with two elements, a vector and a scalar, for the regularization

Method clone()

The objects of this class are cloneable with this method.

Usage

RidgeRegressionFit$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.